rede_social/api/requirements.txt
aniso8601==7.0.0
astroid==2.3.3
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.2.0
django-graphql-jwt==0.3.0
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
promise==2.2.1
ptvsd==4.3.2
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
wrapt==1.11.2
